pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"


[project]
name = "confclasses"
authors = [
  { name="Simon Fletcher", email="simon.fletcher@lexisnexisrisk.com" },
]
description = "A simple wrapper around dataclasses, for general configuration"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "ruamel.yaml"
]
dynamic = ["version"]


[project.license]
file = "LICENCE"


[project.urls]
Homepage = "https://github.com/simonfletcher-ln/confclasses"
Issues = "https://github.com/simonfletcher-ln/confclasses/issues"


[tool.setuptools.dynamic]
version = { attr = "confclasses.__version__" }
